+++ EduMatch/src/entities/SoftSkill.java
@@ -1,0 +1,12 @@
+package entities;
+
+public class SoftSkill extends Game{
+
+    public SoftSkill() {
+    }
+
+    public SoftSkill(int id, String questao, String opcaoCerta, int pontos) {
+        super(id, questao, opcaoCerta, pontos);
+    }
+
+}
